Basic setup for tests, and getting rid of the awkward inheritance.
