+++ packages/runtime-api/package.json
@@ -33,7 +33,7 @@
     "node": "^10.15.3"
   },
   "scripts": {
-    "test": "mocha 'test/**/*.js'",
+    "test": "mocha 'test/**/*.js' --exit",
     "lint": "eslint '**/*.js' --ignore-pattern 'test/**/*.js'"
   },
   "devDependencies": {
